Refactor and massively simpliy related spaces section. fixes #1723
