+++ app/integration_test/tests/sub_spaces.dart
@@ -1,6 +1,6 @@
 import 'package:acter/common/widgets/sliver_scaffold.dart';
 import 'package:acter/features/space/sheets/link_room_sheet.dart';
-import 'package:acter/features/space/widgets/relatest_spaces.dart';
+import 'package:acter/features/space/pages/sub_spaces_page.dart';
 import 'package:convenient_test_dev/convenient_test_dev.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -40,8 +40,8 @@ void subSpaceTests() {
     await t.gotoSpace(spaceId);
     await t.navigateTo([
       const Key('spaces'),
-      RelatedSpaces.moreOptionKey,
-      RelatedSpaces.linkSubspaceKey,
+      SubSpacesPage.moreOptionKey,
+      SubSpacesPage.linkSubspaceKey,
     ]);
 
     final roomListEntry = find.byKey(Key('room-list-link-$subSpace'));
@@ -85,8 +85,8 @@ void subSpaceTests() {
     await t.gotoSpace(spaceId);
     await t.navigateTo([
       const Key('spaces'),
-      RelatedSpaces.moreOptionKey,
-      RelatedSpaces.linkSubspaceKey,
+      SubSpacesPage.moreOptionKey,
+      SubSpacesPage.linkSubspaceKey,
     ]);
 
     final roomListEntry = find.byKey(Key('room-list-link-$subSpace'));
